extract realize_data and OR
