+++ reconcile/test/test_openshift_resource.py
@@ -3,38 +3,31 @@ from .fixtures import Fixtures
 
 import semver
 
-from utils.openshift_resource import OpenshiftResource
+from utils.openshift_resource import OR
 
 fxt = Fixtures('openshift_resource')
 
-QONTRACT_INTEGRATION = 'openshift_resources'
-QONTRACT_INTEGRATION_VERSION = semver.format_version(1, 6, 0)
-
-
-class OR(OpenshiftResource):
-    def __init__(self, body):
-        super(OR, self).__init__(
-            body, QONTRACT_INTEGRATION, QONTRACT_INTEGRATION_VERSION
-        )
+TEST_INT = 'test_openshift_resources'
+TEST_INT_VER = semver.format_version(1, 9, 2)
 
 
 class TestOpenshiftResource(object):
     def test_verify_valid_k8s_object(self):
         resource = fxt.get_anymarkup('valid_resource.yml')
-        openshift_resource = OR(resource)
+        openshift_resource = OR(resource, TEST_INT, TEST_INT_VER)
 
         assert openshift_resource.verify_valid_k8s_object() is None
 
     def test_verify_valid_k8s_object_false(self):
         resource = fxt.get_anymarkup('invalid_resource.yml')
-        openshift_resource = OR(resource)
+        openshift_resource = OR(resource, TEST_INT, TEST_INT_VER)
 
         with pytest.raises(KeyError):
             assert openshift_resource.verify_valid_k8s_object() is None
 
     def test_annotates_resource(self):
         resource = fxt.get_anymarkup('annotates_resource.yml')
-        openshift_resource = OR(resource)
+        openshift_resource = OR(resource, TEST_INT, TEST_INT_VER)
 
         assert openshift_resource.has_qontract_annotations() is False
 
@@ -44,13 +37,15 @@ class TestOpenshiftResource(object):
     def test_sha256sum_properly_ignores_some_params(self):
         resources = fxt.get_anymarkup('ignores_params.yml')
 
-        assert OR(resources[0]).annotate().sha256sum() == \
-            OR(resources[1]).annotate().sha256sum()
+        assert OR(resources[0],
+                  TEST_INT,
+                  TEST_INT_VER).annotate().sha256sum() == \
+            OR(resources[1], TEST_INT, TEST_INT_VER).annotate().sha256sum()
 
     def test_sha256sum(self):
         resource = fxt.get_anymarkup('sha256sum.yml')
 
-        openshift_resource = OR(resource)
+        openshift_resource = OR(resource, TEST_INT, TEST_INT_VER)
 
         assert openshift_resource.sha256sum() == \
             '1366d8ef31f0d83419d25b446e61008b16348b9efee2216873856c49cede6965'
