+++ utils/openshift_resource.py
@@ -195,6 +195,13 @@ class OpenshiftResource(object):
         return m.hexdigest()
 
 
+class OR(OpenshiftResource):
+    def __init__(self, body, integration, integration_version):
+        super(OR, self).__init__(
+            body, integration, integration_version
+        )
+
+
 class ResourceInventory(object):
     def __init__(self):
         self._clusters = {}
